pyproject.toml
[tool.poetry]
name = "mpcover"
version = "0.0.1-alpha1"
description = "Python program for displaying album covers of music currently playing through MPD."
authors = ["Uroš Milivojević <milivojevicu1998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^9.0.1"

[tool.poetry.scripts]
mpcover = 'mpcover.main:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
